Fix sdl2-config behavior
There is a bug in SDL2's CMakeLists.txt which tries to copy libSDL2.so
from an incorrect location. This behavior had previously been patched
out, which had the unfortunate side effect of breaking the behavior of
sdl2-config. The reason it was patched out appears to be because the
CMakeLists.txt does not handle debug-only builds correctly, as it
appends a "d" postfix to the libraries, which trips up both the symlink
and sdl2-config. This behavior has been corrected by both patching out
the debug postfix and correcting the path of the symlink.